Fixed the Data Mapping Json Object response
